vv scripts: update scripts to upload docs
Former-commit-id: 71038d4845a5353305fb858f4d0bd3950beffa91 [formerly 539d8e096df7ed3ac99206e96d7ff623fa97f235]
Former-commit-id: f09f4a527cc7f50975d5b412ead8e6bb2e3ec148
Former-commit-id: 1cfb1f1a5b985eafe70fa45141af5761600e3da0

+++ Utilities/Firebot/smv_guides2GD.sh
@@ -1,5 +1,6 @@
 #!/bin/bash
 
+GDRIVE=~/bin/gdrive
 CURDIR=`pwd`
 # directory containing guides on google drive : FDS-SMV Newest Manuals
 PARENT_ID=0B_wB1pJL2bFQUlJwMmNfaHlqME0
@@ -7,22 +8,21 @@ PARENT_ID=0B_wB1pJL2bFQUlJwMmNfaHlqME0
 # directory containing guides within smokebot account
 FROMDIR=~smokebot/smokebot/output
 BASEDIR=Newest_Guides
-UPDATE_GUIDES=$FROMDIR/$BASEDIR/update_guides
 
 UPLOAD ()
 {
   FILE=$1
   FILEnew=${FILE}_new.pdf
   cp $FILE.pdf /tmp/$FILEnew
-  drive list  | grep $FILEnew | awk '{ system("drive delete -i " $1)} '
-  drive upload -p $PARENT_ID -f /tmp/$FILEnew
+  $GDRIVE list  | grep $FILEnew | awk '{ system("drive delete -i " $1)} '
+  $GDRIVE upload -p $PARENT_ID -f /tmp/$FILEnew
 }
 
-#if [ -e $UPDATE_GUIDES ] ; then
+if [ -e $GDRIVE ] ; then
   cd $FROMDIR/$BASEDIR
   rm -f $UPDATE_GUIDES
   UPLOAD SMV_User_Guide
   UPLOAD SMV_Technical_Reference_Guide
   UPLOAD SMV_Verification_Guide
   cd $CURDIR
-#fi
+fi
